chore: rename apptesting -> testutil
