+++ app/testutil/test_suite.go
@@ -1,4 +1,4 @@
-package apptesting
+package testutil
 
 import (
 	"time"
